Refactor CourtCaseNumber into a component
