--- apps/judicial-system/web/messages/RestrictionCases/Prosecutor/lawsBrokenForm.ts
@@ -1,24 +1,0 @@
-import { defineMessage, defineMessages } from 'react-intl'
-
-// Strings for Laws broken form
-export const lawsBrokenFrom = {
-  sections: {
-    custodyProvisions: defineMessages({
-      title: {
-        id:
-          'judicial.system:forms.restrinction.laws_broken.custody_provision_a-lið1.mgr.95.gr.title',
-        defaultMessage: 'a-lið 1. mgr. 95. gr.',
-        description:
-          'Prosecutor laws broken form. Custody provisions - a-lið 1. mgr. 95. gr.: title',
-      },
-      tooltip: {
-        id:
-          'judicial.system:forms.restrinction.laws_broken.custody_provision_a-lið1.mgr.95.gr.tooltip',
-        defaultMessage:
-          'Að ætla megi að sakborningur muni torvelda rannsókn málsins, svo sem með því að afmá merki eftir brot, skjóta undan munum ellegar hafa áhrif á samseka eða vitni.',
-        description:
-          'Prosecutor laws broken form. Custody provisions - a-lið 1. mgr. 95. gr.: tooltip',
-      },
-    }),
-  },
-}
