+++ apps/judicial-system/web/messages/index.ts
@@ -7,7 +7,6 @@ export * from './requestCourtForm'
 export * from './selectCourt'
 export * from './selectProsecutor'
 export * from './RestrictionCases/Prosecutor/defendantForm'
-export * from './RestrictionCases/Prosecutor/lawsBrokenForm'
 export * from './SharedComponents/policeCaseNumber'
 export * from './SharedComponents/defenderInfo'
 export * from './SharedComponents/requestCourtDate'
