+++ apps/judicial-system/web/src/routes/Court/RestrictionRequest/Overview/Overview.treat.ts
@@ -9,18 +9,3 @@ export const infoSection = style({
   padding: `${theme.spacing[5]}px 0`,
   borderTop: `2px solid ${theme.color.purple100}`,
 })
-
-export const createCourtCaseContainer = style({
-  display: 'flex',
-  flexDirection: 'column',
-})
-
-export const createCourtCaseButton = style({
-  display: 'flex',
-  maxHeight: theme.spacing[8],
-  marginRight: theme.spacing[2],
-})
-
-export const createCourtCaseInput = style({
-  flex: 1,
-})
