+++ apps/judicial-system/web/src/routes/Court/RestrictionRequest/Overview/OverviewForm.tsx
@@ -4,22 +4,15 @@ import {
   Case,
   CaseCustodyProvisions,
   CaseType,
-  IntegratedCourts,
 } from '@island.is/judicial-system/types'
 import {
-  BlueBox,
   CaseFileList,
   FormContentContainer,
   InfoCard,
   PdfButton,
 } from '@island.is/judicial-system-web/src/shared-components'
-import { Box, Button, Input, Text } from '@island.is/island-ui/core'
+import { Box, Button, Text } from '@island.is/island-ui/core'
 import * as styles from './Overview.treat'
-import {
-  removeTabsValidateAndSet,
-  validateAndSendToServer,
-} from '@island.is/judicial-system-web/src/utils/formHelper'
-import { useCase } from '@island.is/judicial-system-web/src/utils/hooks'
 import {
   capitalize,
   formatDate,
@@ -29,6 +22,7 @@ import {
 } from '@island.is/judicial-system/formatters'
 import { UserContext } from '@island.is/judicial-system-web/src/shared-components/UserProvider/UserProvider'
 import { requestCourtDate } from '@island.is/judicial-system-web/messages'
+import CourtCaseNumber from '../../SharedComponents/CourtCaseNumber/CourtCaseNumber'
 
 interface Props {
   workingCase: Case
@@ -54,7 +48,6 @@ const OverviewForm: React.FC<Props> = (props) => {
     setCourtCaseNumberEM,
     setIsDraftingConclusion,
   } = props
-  const { updateCase, isCreatingCourtCase } = useCase()
   const { user } = useContext(UserContext)
   const { formatMessage } = useIntl()
 
@@ -70,79 +63,15 @@ const OverviewForm: React.FC<Props> = (props) => {
         </Text>
       </Box>
       <Box component="section" marginBottom={6}>
-        <Box marginBottom={2}>
-          <Text as="h2" variant="h3">
-            Málsnúmer héraðsdóms
-          </Text>
-        </Box>
-        <Box marginBottom={2}>
-          <Text>
-            Smelltu á hnappinn til að stofna nýtt mál eða skráðu inn málsnúmer
-            sem er þegar til í Auði. Athugið að gögn verða sjálfkrafa vistuð á
-            það málsnúmer sem slegið er inn.
-          </Text>
-        </Box>
-        <BlueBox>
-          <div className={styles.createCourtCaseContainer}>
-            <Box display="flex">
-              {workingCase.court &&
-                IntegratedCourts.includes(workingCase.court.id) && (
-                  <div className={styles.createCourtCaseButton}>
-                    <Button
-                      size="small"
-                      onClick={() => handleCreateCourtCase(workingCase)}
-                      loading={isCreatingCourtCase}
-                      disabled={Boolean(workingCase.courtCaseNumber)}
-                      fluid
-                    >
-                      Stofna nýtt mál
-                    </Button>
-                  </div>
-                )}
-              <div className={styles.createCourtCaseInput}>
-                <Input
-                  data-testid="courtCaseNumber"
-                  name="courtCaseNumber"
-                  label="Mál nr."
-                  placeholder="R-X/ÁÁÁÁ"
-                  size="sm"
-                  backgroundColor="white"
-                  value={workingCase.courtCaseNumber ?? ''}
-                  icon={
-                    workingCase.courtCaseNumber && createCourtCaseSuccess
-                      ? 'checkmark'
-                      : undefined
-                  }
-                  errorMessage={courtCaseNumberEM}
-                  hasError={!isCreatingCourtCase && courtCaseNumberEM !== ''}
-                  onChange={(event) => {
-                    setCreateCourtCaseSuccess(false)
-                    removeTabsValidateAndSet(
-                      'courtCaseNumber',
-                      event,
-                      ['empty'],
-                      workingCase,
-                      setWorkingCase,
-                      courtCaseNumberEM,
-                      setCourtCaseNumberEM,
-                    )
-                  }}
-                  onBlur={(event) => {
-                    validateAndSendToServer(
-                      'courtCaseNumber',
-                      event.target.value,
-                      ['empty'],
-                      workingCase,
-                      updateCase,
-                      setCourtCaseNumberEM,
-                    )
-                  }}
-                  required
-                />
-              </div>
-            </Box>
-          </div>
-        </BlueBox>
+        <CourtCaseNumber
+          workingCase={workingCase}
+          setWorkingCase={setWorkingCase}
+          courtCaseNumberEM={courtCaseNumberEM}
+          setCourtCaseNumberEM={setCourtCaseNumberEM}
+          createCourtCaseSuccess={createCourtCaseSuccess}
+          setCreateCourtCaseSuccess={setCreateCourtCaseSuccess}
+          handleCreateCourtCase={handleCreateCourtCase}
+        />
       </Box>
       <Box component="section" marginBottom={5}>
         <InfoCard
